fix : move map methods for list action to own dedicated file, rename validation files for list actions
